Start working on 3.9: Bug linked to ticket-scenario * REST API - Linking done on create * FRONT - Move the create/update bug form from inline to dialog * FRONT - Update page generation to a more single page application
