--- app/templates/base.html
@@ -1,24 +1,0 @@
-{% extends "base_head.html" %}
-{% block body %}
-<nav class="navbar navbar-expand-lg navbar-light bg-light hstack"
-     id="navigationBar"
-     hx-get="/front/v1/navigation"
-     hx-target="this"
-     hx-swap="innerHTML"
-     hx-trigger="load, navRefresh from:body">
-
-</nav>
-
-<div id="modals-here">
-
-</div>
-
-<div class="container-fluid" id="content-block">
-    {% block contenu %}Mon contenu{% endblock %}
-</div>
-<footer class="sticky-bottom">
-<div class="container" id="messageBox">
-
-</div>
-</footer>
-{% endblock %}
