--- app/templates/base_head.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Dashboard</title>
-    <link id="favicon" rel="icon" type="image/x-icon" href="/assets/favicon.ico">
-    <link href="/assets/bootstrap.min.css" rel="stylesheet">
-    <script src="/assets/bootstrap.bundle.min.js"></script>
-    <script src="/assets/htmx.min.js"></script>
-    <script src="/assets/_hyperscript_web.min.js"></script>
-    <script src="/assets/json-enc.js"></script>
-</head>
-<body>
-{% block body %}The body here{% endblock %}
-</body>
-</html>
