--- app/templates/bug_board.html
@@ -1,7 +1,0 @@
-{% extends "base.html" %}
-{% block titre %}  {% endblock %}
-{% block contenu %}
-<div id="project_{{project_name_alias}}" class="container">
-
-</div>
-{% endblock %}
