+++ app/templates/documentation.html
@@ -1,5 +1,3 @@
-{% extends "base_head.html" %}
-{% block body %}
 
   <div class="container"><div class="hstack"><div class="ms-auto">Application version: {{app_version}}</div> </div> </div>
   <div class="container"
@@ -9,5 +7,3 @@
        id="docContainer">
 
   </div>
-
-{% endblock %}
