+++ app/templates/index.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Dashboard</title>
+    <link id="favicon" rel="icon" type="image/x-icon" href="/assets/favicon.ico">
+    <link href="/assets/bootstrap.min.css" rel="stylesheet">
+    <script src="/assets/bootstrap.bundle.min.js"></script>
+    <script src="/assets/htmx.min.js"></script>
+    <script src="/assets/_hyperscript_web.min.js"></script>
+    <script src="/assets/json-enc.js"></script>
+</head>
+<body>
+    <nav class="navbar navbar-expand-lg navbar-light bg-light hstack"
+     id="navigationBar"
+     hx-get="/front/v1/navigation"
+     hx-target="this"
+     hx-swap="innerHTML"
+     hx-trigger="load, navRefresh from:body">
+    </nav>
+    <div id="modals-here"></div>
+
+    <div class="container-fluid"
+         id="content-block"
+         hx-get="{{path}}"
+         hx-trigger="load">
+
+    </div>
+
+    <footer class="sticky-bottom">
+    <div class="container" id="messageBox">
+
+</div>
+</footer>
+</body>
+</html>
