--- app/templates/project.html
@@ -1,21 +1,0 @@
-{% extends "base_project.html" %}
-<!-- Project data. -->
-{% block project %}
-  <h1>Project: {{project_name}}</h1>
-<!-- project -->
-  <div id="{{project_name_alias}}-versions" class="container">
-    <div id="dialog"></div>
-    <button class="btn btn-primary"
-            hx-get="/front/v1/projects/{{project_name}}/versions"
-            hx-headers='{"eaid-request": "FORM"}'
-            hx-target="#dialog">Add a new version</button>
-    <div id="versionsTable"
-         hx-get="/front/v1/projects/{{project_name}}/versions"
-         hx-trigger="load, update-version-table from:body">
-    </div>
-    <div id="versionTickets">
-    </div>
-    <div id="ticket">
-    </div>
-  </div>
-{% endblock %}
