+++ app/templates/repository_board.html
@@ -1,10 +1,9 @@
-{% extends "base_project.html" %}
-{% block project %}
+
 <div id="project_{{project_name_alias}}" class="container">
     <!-- repository board -->
     <div id="repositoryMessage"></div>
 <div id="{{project_name_alias}}-repository" class="container">
-    <h1>{{project_name}} Repository</h1>
+
     <div id="repository_filter"
          hx-trigger="load"
          hx-get="/front/v1/projects/{{project_name}}/repository/epics-features"
@@ -24,4 +23,3 @@
             >Import and update repository</button>
   </div>
 </div>
-{% endblock %}
