+++ app/templates/test_results.html
@@ -1,5 +1,3 @@
-{% extends "base.html" %}
-{% block contenu %}
 <div class="container" id="results">
 {%for key, result in results.items()%}
     <div class="container" id="result_{{key}}">
@@ -44,4 +42,3 @@
     </div>
 {%endfor%}
 </div>
-{% endblock %}
