+++ app/templates/versions.html
@@ -1,0 +1,15 @@
+ <!-- project -->
+<div id="{{project_name_alias}}-versions" class="container">
+  <div id="dialog"></div>
+  <button class="btn btn-primary"
+          hx-get="/front/v1/projects/{{project_name}}/versions"
+          hx-headers='{"eaid-request": "FORM"}'
+          hx-target="#dialog">Add a new version</button>
+  <div id="versionsTable"
+       hx-get="/front/v1/projects/{{project_name}}/versions"
+       hx-headers='{"eaid-request": "table"}'
+       hx-trigger="load, update-version-table from:body">
+  </div>
+  <div id="versionTickets"></div>
+  <div id="ticket"></div>
+</div>
